Altered the "consistent" parallelization mode for ILU and GaussSeidel
Up to now, this mode used a consistent matrix and additive right-hand sides and treated the correction as additive.
This approach is problematic. It can result in convergence towards a wrong solution even in very simple cases!
The altered implementation uses consistent matrix AND consistent rhs, but the correction is only taken from horizontal masters.
This is achieved by setting the storage type of the correction to PST_UNIQUE (even though it is not)
and then changing to consistent storage type.
This treatment corresponds to the "Restricted Additive Schwarz" method (RAS).